+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+<a name="1.3.1"></a>
+
+## [1.3.1](https://github.wdf.sap.corp/devx-wing/vscode-mta-tools/compare/v1.1.1...v1.3.1) (2020-07-15)
+
+### Bug Fixes
+
+- update command name ([#73](https://github.wdf.sap.corp/devx-wing/vscode-mta-tools/issues/73)) ([90272ad](https://github.wdf.sap.corp/devx-wing/vscode-mta-tools/commit/90272ad))
+
+### Features
+
+- Add command for opening 'add module' wizard ([824baf4](https://github.wdf.sap.corp/devx-wing/vscode-mta-tools/commit/824baf4))
+
+<a name="1.3.0"></a>
+
+# [1.3.0](https://github.wdf.sap.corp/devx-wing/vscode-mta-tools/compare/v1.1.1...v1.3.0) (2020-07-13)
+
+### Features
+
+- Add command for opening 'add module' wizard ([824baf4](https://github.wdf.sap.corp/devx-wing/vscode-mta-tools/commit/824baf4))
+
+<a name="1.1.1"></a>
+
+## [1.1.1](https://github.wdf.sap.corp/devx-wing/vscode-mta-tools/compare/v1.1.0...v1.1.1) (2020-06-23)
+
+### Bug Fixes
+
+- fix space in project name ([#70](https://github.wdf.sap.corp/devx-wing/vscode-mta-tools/issues/70)) ([3552207](https://github.wdf.sap.corp/devx-wing/vscode-mta-tools/commit/3552207))
+
 <a name="1.1.0"></a>
 
 # [1.1.0](https://github.wdf.sap.corp/devx-wing/vscode-mta-tools/compare/v1.0.0...v1.1.0) (2020-02-03)